eliminate extraneous display of "game over" dialog which resulted in re-tweeting the game over if return to line view during halftime
Reported problem: "Right now, there's a pop-up once half-time is reached saying whether we'll receive or defend at the start of the next half.  I guess this pop-up is linked to the auto-tweet function that tweets out "Halftime. [timestamp]" because if you exit out of the app, then come back in, change the line, and go back to the action page, the same pop-up appears and the app auto-tweets again with a new timestamp."